Login

+++ src/main/java/com/gr/comcome/usedBoard/model/usedBoardVO.java
@@ -14,7 +14,7 @@ public class usedBoardVO {
 	private int boardNo;
 	private int accountNo;
 	private String groupNo;
-	private int fileSize;
+	private long fileSize;
 	private String delFlag;
 	private String title;
 	private int readcount;
@@ -27,4 +27,8 @@ public class usedBoardVO {
 	private Timestamp regdate;
 	private String email;
 	private int price;
+	
+	
+	
+	
 }
